graphcap/pipelines/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pipelines"
version = "0.1.0"
description = "Graphcap pipeline package"
requires-python = ">=3.11"
dependencies = [
    "dagster-postgres>=0.25.11",
    "dagster-webserver>=1.9.11",
    "dagster>=1.9.11",
    "datasets>=3.2.0",
    "httpx>=0.28.1",
    "huggingface-hub>=0.28.1",
    "loguru>=0.7.3",
    "openai>=1.61.1",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "pillow>=11.1.0",
    "pyarrow>=19.0.0",
    "tenacity>=9.0.0",
    "tqdm>=4.67.1",
]

[project.optional-dependencies]
dev = [
    "dagster-webserver>=1.9.11",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
]

[tool.dagster]
module_name = "pipelines.definitions"
project_name = "pipelines"

[tool.hatch.build.targets.wheel]
packages = ["pipelines"]

[tool.hatch.build]
packages = ["pipelines"]
include = [
    "pipelines/**/*.py",
]

[tool.uv.sources]
graphcap = { workspace = true }
